Bump minor version and fix some compiler warnings.
git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@6143 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_const_data.cpp
@@ -234,7 +234,7 @@ void spelleffect_data_t::link()
     {
       for( int j=0; spell_data[ j ].name; j++ )
       {
-        if( ed.trigger_spell_id == spell_data[ j ].id )
+        if( ed.trigger_spell_id == (int) spell_data[ j ].id )
         {
           ed.trigger_spell = spell_data + j;
           break;
